Refactoring to align with other components
