--- src/app/admin/admin.service.ts
@@ -1,46 +1,0 @@
-import { Injectable, inject } from '@angular/core';
-import { HttpClient } from '@angular/common/http';
-import { environment } from '../../environments/environment';
-import { UserRole } from '../../api/web-api.types';
-
-export interface User {
-  id: number;
-  email: string;
-  roles: UserRole[];
-}
-
-@Injectable({
-  providedIn: 'root',
-})
-export class AdminService {
-  private http = inject(HttpClient);
-  private baseUrl = `${environment.webApiUrl}/users`;
-
-  getUsers() {
-    return this.http.get<User[]>(this.baseUrl);
-  }
-
-  getUser(id: number) {
-    return this.http.get<User>(`${this.baseUrl}/${id}`);
-  }
-
-  createUser(userData: { email: string; password: string; roles: UserRole[] }) {
-    return this.http.post<{ id: number }>(this.baseUrl, userData);
-  }
-
-  updateUserRoles(userId: number, roles: UserRole[]) {
-    return this.http.put<User>(`${this.baseUrl}/${userId}/roles`, { roles });
-  }
-
-  updatePassword(userId: number, password: string) {
-    return this.http.put(
-      `${this.baseUrl}/${userId}/password`,
-      { password },
-      { responseType: 'text' }
-    );
-  }
-
-  deleteUser(userId: number) {
-    return this.http.delete(`${this.baseUrl}/${userId}`);
-  }
-}
