+++ src/app/admin/user-panel/user-create/create-user.component.html
@@ -1,0 +1,96 @@
+<div class="create-user-dialog">
+  <h2 matDialogTitle>Create a new User</h2>
+
+  <form [formGroup]="userForm" (ngSubmit)="onSubmit()">
+    <mat-dialog-content>
+      <div class="form-fields">
+        <mat-form-field appearance="outline">
+          <mat-label>Email Address</mat-label>
+          <input
+            matInput
+            formControlName="email"
+            type="email"
+            autocomplete="email"
+            [attr.aria-label]="'Email address'"
+          />
+          <mat-error *ngIf="email.errors?.['required']">
+            Email is required
+          </mat-error>
+          <mat-error *ngIf="email.errors?.['email']">
+            Please enter a valid email address
+          </mat-error>
+        </mat-form-field>
+
+        <mat-form-field appearance="outline">
+          <mat-label>Password</mat-label>
+          <input
+            matInput
+            formControlName="password"
+            [type]="hidePassword ? 'password' : 'text'"
+            autocomplete="new-password"
+            [attr.aria-label]="'Password'"
+          />
+          <button
+            mat-icon-button
+            matSuffix
+            type="button"
+            (click)="hidePassword = !hidePassword"
+            [attr.aria-label]="'Toggle password visibility'"
+          >
+            <mat-icon>{{
+              hidePassword ? 'visibility_off' : 'visibility'
+            }}</mat-icon>
+          </button>
+          <mat-error *ngIf="password.errors?.['required']">
+            Password is required
+          </mat-error>
+          <mat-error *ngIf="password.errors?.['minlength']">
+            Password must be at least 8 characters
+          </mat-error>
+        </mat-form-field>
+
+        <mat-form-field appearance="outline">
+          <mat-label>User Roles</mat-label>
+          <mat-select
+            formControlName="roles"
+            multiple
+            [attr.aria-label]="'Select user roles'"
+          >
+            <mat-option *ngFor="let role of availableRoles" [value]="role">
+              {{ role | titlecase }}
+            </mat-option>
+          </mat-select>
+          <mat-error *ngIf="roles.errors?.['required']">
+            Please select at least one role
+          </mat-error>
+        </mat-form-field>
+      </div>
+    </mat-dialog-content>
+
+    <mat-dialog-actions align="end">
+      <button
+        mat-button
+        mat-dialog-close
+        type="button"
+        [disabled]="isSubmitting()"
+      >
+        Cancel
+      </button>
+      <button
+        mat-flat-button
+        color="primary"
+        type="submit"
+        [disabled]="!userForm.valid || isSubmitting()"
+      >
+        <span class="button-content">
+          <mat-spinner
+            *ngIf="isSubmitting()"
+            diameter="20"
+            class="spinner"
+          ></mat-spinner>
+          {{ isSubmitting() ? 'Creating...' : 'Create User' }}
+        </span>
+      </button>
+    </mat-dialog-actions>
+  </form>
+</div>
