+++ src/app/admin/user-panel/user-create/create-user.component.scss
@@ -1,0 +1,51 @@
+.create-user-dialog {
+  mat-dialog-title {
+    margin: 0;
+    padding: 24px 24px 0;
+    font-size: 24px;
+    font-weight: 500;
+    color: rgba(0, 0, 0, 0.87);
+  }
+
+  mat-dialog-content {
+    padding: 24px;
+    margin: 0;
+    min-width: 400px;
+    max-width: 100%;
+    box-sizing: border-box;
+  }
+
+  .form-fields {
+    display: flex;
+    flex-direction: column;
+    gap: 16px;
+  }
+
+  mat-form-field {
+    width: 100%;
+  }
+
+  mat-dialog-actions {
+    padding: 16px 24px;
+    background: #fafafa;
+    margin: 0;
+    border-top: 1px solid rgba(0, 0, 0, 0.12);
+  }
+
+  .button-content {
+    display: flex;
+    align-items: center;
+    gap: 8px;
+  }
+
+  .spinner {
+    margin-right: 8px;
+  }
+
+  @media (max-width: 600px) {
+    mat-dialog-content {
+      min-width: unset;
+      width: 100%;
+    }
+  }
+}
