+++ src/app/admin/user-panel/user-edit-roles/edit-user-roles.component.html
@@ -1,0 +1,30 @@
+    <div class="edit-dialog">
+      <h2 mat-dialog-title>
+        Edit User Roles ({{ data.user?.email ?? 'Unknown' }})
+      </h2>
+
+      <form [formGroup]="userForm" (ngSubmit)="onSubmit()">
+        <mat-dialog-content>
+          <mat-form-field appearance="outline">
+            <mat-label>Roles</mat-label>
+            <mat-select formControlName="roles" multiple>
+              <mat-option *ngFor="let role of availableRoles" [value]="role">
+                {{ role }}
+              </mat-option>
+            </mat-select>
+          </mat-form-field>
+        </mat-dialog-content>
+
+        <mat-dialog-actions align="end" class="dialog-actions">
+          <button mat-button mat-dialog-close>Cancel</button>
+          <button
+            mat-flat-button
+            color="primary"
+            type="submit"
+            [disabled]="!userForm.valid"
+          >
+            Submit
+          </button>
+        </mat-dialog-actions>
+      </form>
+    </div>
