+++ src/app/admin/user-panel/user-edit-roles/edit-user-roles.component.scss
@@ -1,0 +1,29 @@
+.edit-dialog {
+  h2 {
+    margin: 0;
+    padding: 24px 24px 16px;
+    font-size: 24px;
+    font-weight: 500;
+    color: rgba(0, 0, 0, 0.87);
+  }
+}
+
+mat-dialog-content {
+  padding: 0 24px;
+  margin: 0;
+  display: flex;
+  flex-direction: column;
+  gap: 16px;
+  min-width: 400px;
+}
+
+mat-form-field {
+  width: 100%;
+}
+
+.dialog-actions {
+  padding: 16px 24px;
+  background: #fafafa;
+  margin: 0;
+  border-top: 1px solid rgba(0, 0, 0, 0.12);
+}
