+++ src/app/admin/user-panel/user-panel.component.scss
@@ -1,0 +1,78 @@
+.admin-panel {
+  padding: 0;
+
+  h2 {
+    margin: 0;
+    padding: 24px 24px 16px;
+    font-size: 24px;
+    font-weight: 500;
+    color: rgba(0, 0, 0, 0.87);
+  }
+}
+
+.table-container {
+  position: relative;
+  min-height: 200px;
+  max-height: 400px;
+  overflow: auto;
+  background: white;
+}
+
+table {
+  width: 100%;
+  background: white;
+  border-radius: 4px;
+  overflow: hidden;
+
+  th {
+    background: #fafafa;
+    color: rgba(0, 0, 0, 0.87);
+    font-weight: 500;
+    font-size: 14px;
+    padding: 16px;
+  }
+
+  td {
+    padding: 12px 16px;
+    font-size: 14px;
+    color: rgba(0, 0, 0, 0.87);
+  }
+
+  tr:hover {
+    background-color: rgba(0, 0, 0, 0.04);
+  }
+}
+
+.role-chips {
+  display: flex;
+  gap: 8px;
+  flex-wrap: wrap;
+
+  .role-chip {
+    background: #e0e0e0;
+    padding: 4px 12px;
+    border-radius: 16px;
+    font-size: 12px;
+    font-weight: 500;
+    color: rgba(0, 0, 0, 0.87);
+  }
+}
+
+.dialog-actions {
+  padding: 16px 24px;
+  background: #fafafa;
+  margin: 0;
+  border-top: 1px solid rgba(0, 0, 0, 0.12);
+}
+
+mat-dialog-content {
+  padding: 0 24px;
+  margin: 0;
+  max-height: 400px;
+}
+
+.delete-action {
+  .warn-text {
+    color: #f44336;
+  }
+}
