+++ src/app/admin/user-panel/user-update-password/update-password.component.html
@@ -1,0 +1,63 @@
+<div class="update-password-dialog">
+  <h2 matDialogTitle>Update User Password</h2>
+
+  <form [formGroup]="userForm" (ngSubmit)="onSubmit()">
+    <mat-dialog-content>
+      <div class="form-fields">
+        <mat-form-field appearance="outline">
+          <mat-label>New Password</mat-label>
+          <input
+            matInput
+            formControlName="password"
+            [type]="hidePassword ? 'password' : 'text'"
+            autocomplete="new-password"
+            [attr.aria-label]="'Password'"
+          />
+          <button
+            mat-icon-button
+            matSuffix
+            type="button"
+            (click)="hidePassword = !hidePassword"
+            [attr.aria-label]="'Toggle password visibility'"
+          >
+            <mat-icon>{{
+              hidePassword ? 'visibility_off' : 'visibility'
+            }}</mat-icon>
+          </button>
+          <mat-error *ngIf="password.errors?.['required']">
+            Password is required
+          </mat-error>
+          <mat-error *ngIf="password.errors?.['minlength']">
+            Password must be at least 8 characters
+          </mat-error>
+        </mat-form-field>
+      </div>
+    </mat-dialog-content>
+
+    <mat-dialog-actions align="end">
+      <button
+        mat-button
+        mat-dialog-close
+        type="button"
+        [disabled]="isSubmitting()"
+      >
+        Cancel
+      </button>
+      <button
+        mat-flat-button
+        color="primary"
+        type="submit"
+        [disabled]="!userForm.valid || isSubmitting()"
+      >
+        <span class="button-content">
+          <mat-spinner
+            *ngIf="isSubmitting()"
+            diameter="20"
+            class="spinner"
+          ></mat-spinner>
+          {{ isSubmitting() ? 'Updating...' : 'Submit' }}
+        </span>
+      </button>
+    </mat-dialog-actions>
+  </form>
+</div>
